pyproject.toml
[project]
name = "mlex_pytorch_autoencoders"
version = "0.2.1"
description = ""
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dvclive[lightning]<3.0",
    "dvc-objects<=1.3.2",
    "einops==0.7.0",
    "numpy<2.0.0",
    "pandas==2.1.1",
    "Pillow>=10.2.0",
    "pyarrow>=14.0.1",
    "pydantic==1.8.2",
    "pygit2==1.14.1",
    "pytorch-lightning==1.8.6",
    "tifffile",
    "tiled[client]==0.1.0a118",
    "torchvision==0.17.2",
    "pyFAI==2023.9.0",
    "pyyaml",
    "mlflow",
    "numpy<2"
]

[tool.setuptools.packages]
find = { where = ["src"] }

[project.optional-dependencies]
# These are required for developing the package (running the tests, building
# the documentation) but not necessarily required for _using_ it.
dev = [
    "black==24.2.0",
    "flake8==7.0.0",
    "pre-commit==3.6.2",
    "isort==5.13.2",
    "pytest"
]

[tool.isort]
profile = "black"
